src/test/java/com/roamer/contacts/controller/UserControllerTest.java
package com.roamer.contacts.controller;

import com.roamer.contacts.entity.User;
import com.roamer.contacts.service.UserService;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.MockitoAnnotations;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

@RunWith(SpringRunner.class)
@SpringBootTest
public class UserControllerTest {

    @InjectMocks
    private UserController userController;

    @Mock
    private UserService userService;

    private MockMvc mockMvc;

    @Before
    public void setUp() throws Exception {
        //注入所需属性
        MockitoAnnotations.initMocks(this);
        mockMvc = MockMvcBuilders.standaloneSetup(userController).build();
    }

    @Test
    public void index() throws Exception {
        mockMvc.perform(MockMvcRequestBuilders.get("/user/index"))
                .andExpect(MockMvcResultMatchers.status().isOk())
                .andExpect(MockMvcResultMatchers.view().name("user/index"))
                .andDo(MockMvcResultHandlers.print())
                .andReturn();
    }

    @Test
    public void show() throws Exception {
        User user = new User();
        user.setId(1);
        user.setUsername("jbauer");
        user.setPassword("24hours");
        Mockito.when(userService.findUserByUsername("jbauer")).thenReturn(user);
        mockMvc.perform(
                MockMvcRequestBuilders.get("/user/show")
                        .param("username","jbauer").accept(MediaType.APPLICATION_JSON))
                .andExpect(MockMvcResultMatchers.status().isOk())
                .andExpect(MockMvcResultMatchers.content().contentType(MediaType.APPLICATION_JSON_UTF8_VALUE))
                .andExpect(MockMvcResultMatchers.jsonPath("$.id").value(1))
                .andExpect(MockMvcResultMatchers.jsonPath("$.username").value("jbauer"))
                .andExpect(MockMvcResultMatchers.jsonPath("$.password").exists())
                .andDo(MockMvcResultHandlers.print());


    }
}